Migrate the LAST and final boss X86 plugin ##arch
* anal.cc tests are the only tests marked as broken
* Bonus points: minor code cleanup here and there
* Thanks to Luc for all the work and chats to make it happen
