+++ libr/anal/fcn.c
@@ -1598,7 +1598,8 @@ R_API void r_anal_trim_jmprefs(RAnal *anal, RAnalFunction *fcn) {
 	RAnalRef *ref;
 	RList *refs = r_anal_function_get_refs (fcn);
 	RListIter *iter;
-	const bool is_x86 = anal->cur->arch && !strcmp (anal->cur->arch, "x86"); // HACK
+	const char *arch = R_UNWRAP4 (anal, arch, session, name);
+	const bool is_x86 = arch && !strcmp (arch, "x86"); // HACK
 
 	r_list_foreach (refs, iter, ref) {
 		int rt = R_ANAL_REF_TYPE_MASK (ref->type);
@@ -1782,12 +1783,9 @@ R_API bool r_anal_function_add_bb(RAnal *a, RAnalFunction *fcn, ut64 addr, ut64 
 		block = NULL;
 	}
 
-	const char *sarch = R_UNWRAP3 (a, cur, arch);
-	if (!sarch && a->arch->cfg) {
-		sarch = a->arch->cfg->arch;
-	}
-	const bool is_x86 = sarch && !strcmp (sarch, "x86");
-	// TODO fix this x86-ism
+	// XXX R2_592 - try to remove this check, no need to be x86 specific
+	const char *sarch = R_UNWRAP5 (a, arch, session, config, arch);
+	const bool is_x86 = sarch && r_str_startswith (sarch, "x86");
 	if (is_x86) {
 		fcn_recurse (a, fcn, addr, size, -1);
 		block = r_anal_get_block_at (a, addr);
@@ -2063,12 +2061,14 @@ R_API RAnalBlock *r_anal_function_bbget_in(RAnal *anal, RAnalFunction *fcn, ut64
 	}
 	RListIter *iter;
 	RAnalBlock *bb;
-	const bool jmpmid = r_anal_is_aligned (anal, addr);
+	const bool aligned = r_anal_is_aligned (anal, addr);
 	r_list_foreach (fcn->bbs, iter, bb) {
 		if (r_anal_block_contains (bb, addr)) {
-			if ((!anal->opt.jmpmid || !jmpmid || r_anal_block_op_starts_at (bb, addr))) {
+			if ((!anal->opt.jmpmid || !aligned || r_anal_block_op_starts_at (bb, addr))) {
+			// if (r_anal_block_op_starts_at (bb, addr)) {
 				return bb;
 			}
+			// return bb;
 		}
 	}
 	return NULL;
