+++ libr/anal/meson.build
@@ -39,49 +39,7 @@ r_anal_sources = [
   'var.c',
   'vtable.c',
   'xrefs.c',
-  '../arch/p/6502_cs/plugin.c',
-  '../arch/p/avr/plugin.c',
-  '../arch/p/avr/avr_disasm.c',
-  '../arch/p/avr/format.c',
-  '../arch/p/avr/disasm.c',
-  '../arch/p/avr/assemble.c',
-  '../arch/p/ebc/plugin.c',
-  '../arch/p/ebc/ebc_disas.c',
-  '../arch/p/gb/plugin.c',
-  '../arch/p/h8300/plugin.c',
-  '../arch/p/h8300/h8300_disas.c',
-  '../arch/p/lh5801/plugin.c',
-  '../arch/p/m68k_cs/plugin.c',
-  '../arch/p/m680x_cs/plugin.c',
-  '../arch/p/m68k_gnu/plugin.c',
-  '../arch/p/m68k_gnu/m68k-dis.c',
-  '../arch/p/m68k_gnu/m68k-opc.c',
-  '../arch/p/loongarch/plugin.c',
-  '../arch/p/loongarch/loongarch-coder.c',
-  '../arch/p/loongarch/loongarch-dis.c',
-  '../arch/p/loongarch/loongarch-opc.c',
   'p/anal_null.c',
-  'p/anal_x86_cs.c',
-  '../arch/p/xcore_cs/plugin.c',
-  '../arch/p/msp430/plugin.c',
-  '../arch/p/msp430/msp430_disas.c',
-  '../arch/p/arm/winedbg/be_arm.c',
-  '../arch/p/arm/gnu/floatformat.c',
-  '../arch/p/arm/gnu/arm-dis.c',
-  '../arch/p/arm/aarch64/aarch64-dis.c',
-  '../arch/p/arm/aarch64/aarch64-dis-2.c',
-  '../arch/p/arm/aarch64/aarch64-opc.c',
-  '../arch/p/arm/aarch64/aarch64-opc-2.c',
-  '../arch/p/ppc/libvle/vle.c',
-  '../arch/p/ppc/libps/libps.c',
-  '../arch/p/tms320/plugin.c',
-  '../arch/p/tms320/tms320_dasm.c',
-  '../arch/p/tms320/c55x_plus/ins.c',
-  '../arch/p/tms320/c55x_plus/c55plus.c',
-  '../arch/p/tms320/c55x_plus/c55plus_decode.c',
-  '../arch/p/tms320/c55x_plus/decode_funcs.c',
-  '../arch/p/tms320/c55x_plus/hashtable.c',
-  '../arch/p/tms320/c55x_plus/utils.c',
 ]
 
 r_anal_sources += r_anal_d_sources
