+++ libr/arch/p/Makefile
@@ -1,6 +1,4 @@
 include ../../config.mk
-include ../../../mk/platform.mk
-
 ifeq ($(WITHPIC),1)
 CFLAGS+=-I../../include -I../arch/ -I../arch/include -Wall ${PIC_CFLAGS} ${LDFLAGS_LIB}
 CFLAGS+=-DR2__UNIX__
@@ -9,8 +7,8 @@ all: ${ALL_TARGETS}
 
 ALL_TARGETS=
 #x86.mk arm.mk
-ARCHS=null.mk i4004.mk amd29k.mk
-include $(ARCHS)
+#ARCHS=null.mk i4004.mk amd29k.mk
+#include $(ARCHS)
 
 clean:
 	-rm -f *.so *.o ${STATIC_OBJ}
@@ -24,3 +22,4 @@ all clean mrproper:
 
 .PHONY: all clean mrproper
 endif
+include ../../../mk/platform.mk
