+++ libr/include/r_arch.h
@@ -260,85 +260,86 @@ R_API const char *r_arch_op_family_tostring(int n);
 R_API int r_arch_op_family_from_string(const char *f);
 R_API const char *r_arch_op_direction_tostring(struct r_anal_op_t *op);
 
-extern const RArchPlugin r_arch_plugin_null;
-extern const RArchPlugin r_arch_plugin_i4004;
+extern const RArchPlugin r_arch_plugin_6502;
+extern const RArchPlugin r_arch_plugin_6502_cs;
+extern const RArchPlugin r_arch_plugin_8051;
+extern const RArchPlugin r_arch_plugin_alpha;
 extern const RArchPlugin r_arch_plugin_amd29k;
-extern const RArchPlugin r_arch_plugin_jdh8;
-extern const RArchPlugin r_arch_plugin_kvx;
-extern const RArchPlugin r_arch_plugin_pickle;
-extern const RArchPlugin r_arch_plugin_sh;
-extern const RArchPlugin r_arch_plugin_sh_cs;
-extern const RArchPlugin r_arch_plugin_v810;
-extern const RArchPlugin r_arch_plugin_rsp;
-extern const RArchPlugin r_arch_plugin_riscv;
-extern const RArchPlugin r_arch_plugin_riscv_cs;
 extern const RArchPlugin r_arch_plugin_any_as;
 extern const RArchPlugin r_arch_plugin_any_vasm;
+extern const RArchPlugin r_arch_plugin_arc;
 extern const RArchPlugin r_arch_plugin_arm;
-extern const RArchPlugin r_arch_plugin_wasm;
-extern const RArchPlugin r_arch_plugin_x86_nz;
-extern const RArchPlugin r_arch_plugin_x86_nasm;
-extern const RArchPlugin r_arch_plugin_snes;
-extern const RArchPlugin r_arch_plugin_6502;
-extern const RArchPlugin r_arch_plugin_xap;
-extern const RArchPlugin r_arch_plugin_v850;
-extern const RArchPlugin r_arch_plugin_propeller;
-extern const RArchPlugin r_arch_plugin_mcore;
-extern const RArchPlugin r_arch_plugin_nios2;
-extern const RArchPlugin r_arch_plugin_xtensa;
-extern const RArchPlugin r_arch_plugin_or1k;
-extern const RArchPlugin r_arch_plugin_evm;
-extern const RArchPlugin r_arch_plugin_dis;
-extern const RArchPlugin r_arch_plugin_mcs96;
-extern const RArchPlugin r_arch_plugin_ws;
-extern const RArchPlugin r_arch_plugin_lanai;
-extern const RArchPlugin r_arch_plugin_lua;
-extern const RArchPlugin r_arch_plugin_z80;
-extern const RArchPlugin r_arch_plugin_lm32;
+extern const RArchPlugin r_arch_plugin_arm_cs;
+extern const RArchPlugin r_arch_plugin_arm_gnu;
+extern const RArchPlugin r_arch_plugin_arm_v35;
+extern const RArchPlugin r_arch_plugin_avr;
+extern const RArchPlugin r_arch_plugin_bf;
 extern const RArchPlugin r_arch_plugin_bpf;
 extern const RArchPlugin r_arch_plugin_bpf_cs;
-extern const RArchPlugin r_arch_plugin_alpha;
-extern const RArchPlugin r_arch_plugin_vax;
-extern const RArchPlugin r_arch_plugin_tricore;
-extern const RArchPlugin r_arch_plugin_tricore_cs;
-extern const RArchPlugin r_arch_plugin_pic;
-extern const RArchPlugin r_arch_plugin_arm_v35;
-extern const RArchPlugin r_arch_plugin_cris;
-extern const RArchPlugin r_arch_plugin_cr16;
-extern const RArchPlugin r_arch_plugin_arc;
-extern const RArchPlugin r_arch_plugin_pdp11;
-extern const RArchPlugin r_arch_plugin_lh5801;
-extern const RArchPlugin r_arch_plugin_ebc;
-extern const RArchPlugin r_arch_plugin_msp430;
-extern const RArchPlugin r_arch_plugin_pyc;
-extern const RArchPlugin r_arch_plugin_h8300;
-extern const RArchPlugin r_arch_plugin_bf;
-extern const RArchPlugin r_arch_plugin_sparc_gnu;
-extern const RArchPlugin r_arch_plugin_sparc_cs;
-extern const RArchPlugin r_arch_plugin_hppa_gnu;
-extern const RArchPlugin r_arch_plugin_s390_cs;
-extern const RArchPlugin r_arch_plugin_s390_gnu;
-extern const RArchPlugin r_arch_plugin_m68k_gnu;
-extern const RArchPlugin r_arch_plugin_m68k_cs;
-extern const RArchPlugin r_arch_plugin_ppc_gnu;
-extern const RArchPlugin r_arch_plugin_loongarch_gnu;
-extern const RArchPlugin r_arch_plugin_6502_cs;
-extern const RArchPlugin r_arch_plugin_m680x_cs;
-extern const RArchPlugin r_arch_plugin_xcore_cs;
 extern const RArchPlugin r_arch_plugin_chip8;
-extern const RArchPlugin r_arch_plugin_mips_gnu;
-extern const RArchPlugin r_arch_plugin_sm5xx;
-extern const RArchPlugin r_arch_plugin_tms320;
-extern const RArchPlugin r_arch_plugin_ppc_cs;
+extern const RArchPlugin r_arch_plugin_cr16;
+extern const RArchPlugin r_arch_plugin_cris;
+extern const RArchPlugin r_arch_plugin_dalvik;
+extern const RArchPlugin r_arch_plugin_dis;
+extern const RArchPlugin r_arch_plugin_ebc;
+extern const RArchPlugin r_arch_plugin_evm;
+extern const RArchPlugin r_arch_plugin_gb;
+extern const RArchPlugin r_arch_plugin_h8300;
+extern const RArchPlugin r_arch_plugin_hppa_gnu;
+extern const RArchPlugin r_arch_plugin_i4004;
 extern const RArchPlugin r_arch_plugin_i8080;
 extern const RArchPlugin r_arch_plugin_java;
-extern const RArchPlugin r_arch_plugin_arm_gnu;
-extern const RArchPlugin r_arch_plugin_arm_cs;
-extern const RArchPlugin r_arch_plugin_gb;
+extern const RArchPlugin r_arch_plugin_jdh8;
+extern const RArchPlugin r_arch_plugin_kvx;
+extern const RArchPlugin r_arch_plugin_lanai;
+extern const RArchPlugin r_arch_plugin_lh5801;
+extern const RArchPlugin r_arch_plugin_lm32;
+extern const RArchPlugin r_arch_plugin_loongarch_gnu;
+extern const RArchPlugin r_arch_plugin_lua;
+extern const RArchPlugin r_arch_plugin_m680x_cs;
+extern const RArchPlugin r_arch_plugin_m68k_cs;
+extern const RArchPlugin r_arch_plugin_m68k_gnu;
+extern const RArchPlugin r_arch_plugin_mcore;
+extern const RArchPlugin r_arch_plugin_mcs96;
 extern const RArchPlugin r_arch_plugin_mips_cs;
-extern const RArchPlugin r_arch_plugin_avr;
-extern const RArchPlugin r_arch_plugin_8051;
-extern const RArchPlugin r_arch_plugin_dalvik;
+extern const RArchPlugin r_arch_plugin_mips_gnu;
+extern const RArchPlugin r_arch_plugin_msp430;
+extern const RArchPlugin r_arch_plugin_nios2;
+extern const RArchPlugin r_arch_plugin_null;
+extern const RArchPlugin r_arch_plugin_or1k;
+extern const RArchPlugin r_arch_plugin_pdp11;
+extern const RArchPlugin r_arch_plugin_pic;
+extern const RArchPlugin r_arch_plugin_pickle;
+extern const RArchPlugin r_arch_plugin_ppc_cs;
+extern const RArchPlugin r_arch_plugin_ppc_gnu;
+extern const RArchPlugin r_arch_plugin_propeller;
+extern const RArchPlugin r_arch_plugin_pyc;
+extern const RArchPlugin r_arch_plugin_riscv;
+extern const RArchPlugin r_arch_plugin_riscv_cs;
+extern const RArchPlugin r_arch_plugin_rsp;
+extern const RArchPlugin r_arch_plugin_s390_cs;
+extern const RArchPlugin r_arch_plugin_s390_gnu;
+extern const RArchPlugin r_arch_plugin_sh;
+extern const RArchPlugin r_arch_plugin_sh_cs;
+extern const RArchPlugin r_arch_plugin_sm5xx;
+extern const RArchPlugin r_arch_plugin_snes;
+extern const RArchPlugin r_arch_plugin_sparc_cs;
+extern const RArchPlugin r_arch_plugin_sparc_gnu;
+extern const RArchPlugin r_arch_plugin_tms320;
+extern const RArchPlugin r_arch_plugin_tricore;
+extern const RArchPlugin r_arch_plugin_tricore_cs;
+extern const RArchPlugin r_arch_plugin_v810;
+extern const RArchPlugin r_arch_plugin_v850;
+extern const RArchPlugin r_arch_plugin_vax;
+extern const RArchPlugin r_arch_plugin_wasm;
+extern const RArchPlugin r_arch_plugin_ws;
+extern const RArchPlugin r_arch_plugin_x86_cs;
+extern const RArchPlugin r_arch_plugin_x86_nasm;
+extern const RArchPlugin r_arch_plugin_x86_nz;
+extern const RArchPlugin r_arch_plugin_xap;
+extern const RArchPlugin r_arch_plugin_xcore_cs;
+extern const RArchPlugin r_arch_plugin_xtensa;
+extern const RArchPlugin r_arch_plugin_z80;
 
 #ifdef __cplusplus
 }
