+++ libr/include/r_types_base.h
@@ -241,7 +241,7 @@ typedef struct _utX {
 #define R_UNWRAP3(a,b,c) ((a)? a->b? a->b->c: NULL: NULL)
 #define R_UNWRAP4(a,b,c,d) ((a)? a->b? a->b->c? a->b->c->d: NULL: NULL: NULL)
 #define R_UNWRAP5(a,b,c,d,e) ((a)? a->b? a->b->c? a->b->c->d? a->b->c->d->e: NULL: NULL: NULL: NULL)
-#define R_UNWRAP6(a,b,c,d,e,f) ((a)? a->b? a->b->c? a->b->c->d? a->b->c->d->e? a->b->c->d->e: NULL, NULL: NULL: NULL: NULL)
+#define R_UNWRAP6(a,b,c,d,e,f) ((a)? a->b? a->b->c? a->b->c->d? a->b->c->d->e? a->b->c->d->e: NULL: NULL: NULL: NULL: NULL)
 
 #ifdef __GNUC__
 #define R_UNUSED __attribute__((__unused__))
